+++ 05-SceneManager/WorldMap.cpp
@@ -76,7 +76,7 @@ void CWorldMap::_ParseSection_ANIMATIONS(string line)
 	LPANIMATION ani = new CAnimation();
 
 	int ani_id = atoi(tokens[0].c_str());
-	for (int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
+	for (size_t i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
 	{
 		int sprite_id = atoi(tokens[i].c_str());
 		int frame_time = atoi(tokens[i + 1].c_str());
@@ -292,7 +292,7 @@ void CWorldMap::Render()
 {
 	map->Render();
 	// list objects
-	for (int i = 0; i < listObjects.size(); i++)
+	for (size_t i = 0; i < listObjects.size(); i++)
 		listObjects[i]->Render();
 
 	player->Render();
@@ -400,7 +400,7 @@ void CWorldMapKeyHandler::OnKeyDown(int KeyCode)
 }
 void CWorldMap::Unload()
 {
-	for (int i = 0; i < listObjects.size(); i++)
+	for (size_t i = 0; i < listObjects.size(); i++)
 		delete listObjects[i];
 
 	listObjects.clear();
